Get winter fuel payment threshold from rate file
1. Get the new winter fuel payment rate from the YAML file.

The date of birth threshold for the 2016/2017 year should be set to
5th May 1953, as stated here: https://www.gov.uk/winter-fuel-payment.

2. Update winter fuel payment eligibility tests

i. Remove the check for winter fuel payment eligibility from the integration
tests. This should be check in the unit tests.

ii. Update the winter fuel eligibility unit test to read the winter fuel
threshold date from the rate file rather than hard coding the date, to
make the test more generic.

iii. Update the integration tests to account for new winter fuel payment
threshold date. The threshold date is likely to be updated every year
or so. To avoid having to update the tests every time the threshold
changed, time travel to a point in time where the date of birth
entered is below the winter fuel payment threshold, and for the purposes
of the test, not eligible.