Generate expected results for 1st May 1953 Update the expected results for a date of birth of 1st May 1953
Update winter_fuel_payment unit test. Get the 
winter_fuel_payment_threshold rate and use it in the test.

Update integration tests an remove the checks for the value of
age_variant.

The age_variant function is being tested
in the unit tests, it doesn't need to be re-tested in the integration
tests